+++ src/main/java/info/ata4/minecraft/minema/Minema.java
@@ -11,7 +11,6 @@ package info.ata4.minecraft.minema;
 
 import java.io.File;
 
-import info.ata4.minecraft.minema.client.KeyHandler;
 import info.ata4.minecraft.minema.client.cmd.CommandMinema;
 import info.ata4.minecraft.minema.client.config.MinemaConfig;
 import info.ata4.minecraft.minema.client.modules.CaptureSession;
